Extract reusable config function from `Sidebar`
Will be reused by `Notebook`
