+++ src/annotator/sidebar.js
@@ -2,6 +2,7 @@ import Hammer from 'hammerjs';
 
 import annotationCounts from './annotation-counts';
 import sidebarTrigger from './sidebar-trigger';
+import { createSidebarConfig } from './config/sidebar';
 import events from '../shared/bridge-events';
 import features from './features';
 
@@ -23,46 +24,6 @@ const defaultConfig = {
     container: '.annotator-frame',
   },
 };
-
-/**
- * Create the JSON-serializable subset of annotator configuration that should
- * be passed to the sidebar application.
- */
-function createSidebarConfig(config) {
-  const sidebarConfig = { ...config };
-
-  // Some config settings are not JSON-stringifiable (e.g. JavaScript
-  // functions) and will be omitted when the config is JSON-stringified.
-  // Add a JSON-stringifiable option for each of these so that the sidebar can
-  // at least know whether the callback functions were provided or not.
-  if (sidebarConfig.services?.length > 0) {
-    const service = sidebarConfig.services[0];
-    if (service.onLoginRequest) {
-      service.onLoginRequestProvided = true;
-    }
-    if (service.onLogoutRequest) {
-      service.onLogoutRequestProvided = true;
-    }
-    if (service.onSignupRequest) {
-      service.onSignupRequestProvided = true;
-    }
-    if (service.onProfileRequest) {
-      service.onProfileRequestProvided = true;
-    }
-    if (service.onHelpRequest) {
-      service.onHelpRequestProvided = true;
-    }
-  }
-
-  // Remove several annotator-only properties.
-  //
-  // nb. We don't currently strip all the annotator-only properties here.
-  // That's OK because validation / filtering happens in the sidebar app itself.
-  // It just results in unnecessary content in the sidebar iframe's URL string.
-  ['sidebarAppUrl', 'pluginClasses'].forEach(key => delete sidebarConfig[key]);
-
-  return sidebarConfig;
-}
 
 /**
  * Create the iframe that will load the sidebar application.
